修正 "編輯權限API" Bug 修正 "刪除權限API" 刪除前先檢查是否跟角色榜定

+++ Application/src/main/java/com/shark/application/service/permission/DeletePermissionService.java
@@ -1,6 +1,9 @@
 package com.shark.application.service.permission;
 
+import com.shark.application.exception.ResponseException;
 import com.shark.application.repository.permission.PermissionRepository;
+import com.shark.application.repository.role.RolePermissionRepository;
+import com.shark.application.repository.role.dao.RolePermissionDaoEntity;
 import com.shark.application.service.BaseResponseService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -17,6 +20,9 @@ public class DeletePermissionService extends BaseResponseService {
     public static final String INPUT_ID = "id";
 
     @Autowired
+    private RolePermissionRepository rolePermissionRepository;
+
+    @Autowired
     private PermissionRepository permissionRepository;
 
     @Override
@@ -29,6 +35,10 @@ public class DeletePermissionService extends BaseResponseService {
     @Override
     protected Void dataAccess(HashMap<String, String> parameters) {
         long id = Long.valueOf(parameters.get(INPUT_ID));
+        RolePermissionDaoEntity rolePermissionDaoEntity = rolePermissionRepository.findByPermissionId(id);
+        if(rolePermissionDaoEntity != null) {
+            throw new ResponseException(-1, "權限跟角色榜定，請先刪除榜定");
+        }
         permissionRepository.delete(id);
         return null;
     }
